+++ doc-user-mkdocs-project/docs/appendix-release-notes/release-notes.md
@@ -7,6 +7,8 @@ See the [TSTool release notes](http://opencdss.state.co.us/tstool/latest/doc-use
 Plugin release notes are listed below.
 The repository issue for release note item is shown where applicable.
 
+*   [Version 1.5.4](#version-154)
+*   [Version 1.5.3](#version-153)
 *   [Version 1.5.2](#version-152)
 *   [Version 1.5.1](#version-151)
 *   [Version 1.5.0](#version-150)
@@ -21,12 +23,19 @@ The repository issue for release note item is shown where applicable.
 
 ----------
 
+## Version 1.5.4 ##
+
+**Maintenance release to improve the [`AwsBilling`](../command-ref/AwsBilling/AwsBilling.md) command.**
+
+*   ![change](change.png) Update the [`AwsBilling`](../command-ref/AwsBilling/AwsBilling.md) command:
+    +   [#50] Fix VPN connection and Elastic IP service properties table output to show tags.
+
 ## Version 1.5.3 ##
 
 **Maintenance release to improve the [`AwsBilling`](../command-ref/AwsBilling/AwsBilling.md) command.**
 
 *   ![change](change.png) Update the [`AwsBilling`](../command-ref/AwsBilling/AwsBilling.md) command:
-    +   [#50] Add VPN and Elastic IP services to the service properties table.
+    +   [#50] Add VPN connection and Elastic IP services to the service properties table.
     +   [#50] Add public and private IP and DNS name output to the service properties table.
 
 ## Version 1.5.2 ##